# OpenRestoApi/Controllers/AdminController.cs
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using OpenRestoApi.Core.Application.DTOs;
using OpenRestoApi.Core.Application.Exceptions;
using OpenRestoApi.Core.Application.Services;
using OpenRestoApi.Core.Application.Utilities;
using OpenRestoApi.Infrastructure.Auth;

namespace OpenRestoApi.Controllers;

[ApiController]
[Route("api/[controller]")]
[Authorize(Policy = AuthPolicies.RequireAdmin)]
public class AdminController(AdminService adminService) : ControllerBase
{
    public enum bookingStatus { active, cancelled, all, past, upcoming }
    private readonly AdminService _adminService = adminService;

    // Aggregates counts/lists across restaurants and bookings; gated on the dominant resource
    // (bookings — TodayBookingsList and the booking counts/occupancy chart are the bulk of the
    // payload) rather than split across two scopes for one read. A key without bookings:read
    // cannot see the dashboard at all; TodayBookingsList is still guest-redacted independently
    // by BookingGuestVisibility when the key also lacks guests:read.
    [HttpGet("overview")]
    [RequiresScope(ApiKeyScopes.Bookings, ApiKeyScopes.Read)]
    public async Task<IActionResult> Overview()
        => Ok(await _adminService.GetOverviewAsync());

    [HttpGet("bookings")]
    [RequiresScope(ApiKeyScopes.Bookings, ApiKeyScopes.Read)]
    public async Task<IActionResult> GetBookings(
        [FromQuery] int? restaurantId,
        [FromQuery] DateTime? date,
        [FromQuery] bookingStatus status = bookingStatus.active,
        [FromQuery] bool cancelled = false,
        [FromQuery] string? email = null,
        [FromQuery] string? bookingRef = null,
        [FromQuery] string? query = null)
    {
        //this is business logic that should likely belong in the service but its OK for now
        string effectiveStatus = cancelled ? nameof(bookingStatus.cancelled) : status.ToString();
        return Ok(await _adminService.GetBookingsAsync(restaurantId, date, effectiveStatus, email, bookingRef, query));
    }

    [HttpGet("bookings/{id}")]
    [RequiresScope(ApiKeyScopes.Bookings, ApiKeyScopes.Read)]
    public async Task<IActionResult> GetBooking(int id)
    {
        BookingDetailDto? result = await _adminService.GetBookingAsync(id);
        return result == null ? NotFound() : Ok(result);
    }

    [HttpPost("bookings")]
    [RequiresScope(ApiKeyScopes.Bookings, ApiKeyScopes.Write)]
    public async Task<IActionResult> CreateBooking([FromBody] AdminCreateBookingRequest req)
    {
        // ValidationException (bad table/section) → 400, ConflictException (overlap/seats) → 409
        // are mapped by GlobalExceptionHandler; the controller just orchestrates.
        BookingDetailDto result = await _adminService.CreateBookingAsync(req);
        return CreatedAtAction(nameof(GetBooking), new { id = result.Id }, result);
    }

    [HttpPost("bookings/{id}/extend")]
    [RequiresScope(ApiKeyScopes.Bookings, ApiKeyScopes.Write)]
    public async Task<IActionResult> ExtendBooking(int id, [FromBody] ExtendBookingRequest req)
    {
        DateTime? endTime = await _adminService.ExtendBookingAsync(id, req.Minutes);
        return endTime == null ? NotFound() : Ok(new { endTime });
    }

    [HttpPost("bookings/{id}/cancel")]
    [RequiresScope(ApiKeyScopes.Bookings, ApiKeyScopes.Write)]
    public async Task<IActionResult> CancelBooking(int id)
    {
        // ConflictException (past booking) → 409 is mapped by GlobalExceptionHandler.
        return await _adminService.CancelBookingAsync(id) ? NoContent() : NotFound();
    }

    [HttpDelete("bookings/{id}")]
    [RequiresScope(ApiKeyScopes.Bookings, ApiKeyScopes.Write)]
    public async Task<IActionResult> PurgeBooking(int id)
        => await _adminService.PurgeBookingAsync(id) ? NoContent() : NotFound();

    [HttpPost("restaurants")]
    [RequiresScope(ApiKeyScopes.Locations, ApiKeyScopes.Write)]
    public async Task<IActionResult> CreateRestaurant([FromBody] CreateRestaurantRequest req)
    {
        if (string.IsNullOrWhiteSpace(req.Name))
        {
            throw new ValidationException("Name is required.") { Code = ErrorCodes.RestaurantNameRequired };
        }

        RestaurantDto result = await _adminService.CreateRestaurantAsync(req.Name, req.Address);
        return CreatedAtAction(nameof(Overview), new { }, result);
    }

    [HttpPatch("restaurants/{id}")]
    [RequiresScope(ApiKeyScopes.Locations, ApiKeyScopes.Write)]
    public async Task<IActionResult> PatchRestaurant(int id, [FromBody] AdminRestaurantPatchRequest req)
    {
        if (req.IsArchived.HasValue)
        {
            bool success = await _adminService.SetArchivedAsync(id, req.IsArchived.Value);
            if (!success) return NotFound();
        }
        return NoContent();
    }

    [HttpGet("restaurants/{id}/delete-preview")]
    [Authorize(Policy = AuthPolicies.RequireOwner)]
    [RequiresScope(ApiKeyScopes.Locations, ApiKeyScopes.Read)]
    public async Task<IActionResult> GetRestaurantDeletePreview(int id)
    {
        RestaurantDeletePreviewDto? preview = await _adminService.GetRestaurantDeletePreviewAsync(id);
        return preview == null ? NotFound() : Ok(preview);
    }

    // Owner-only: the one irreversible cascade in the admin, gated like user management.
    [HttpDelete("restaurants/{id}")]
    [Authorize(Policy = AuthPolicies.RequireOwner)]
    [RequiresScope(ApiKeyScopes.Locations, ApiKeyScopes.Write)]
    public async Task<IActionResult> DeleteRestaurant(int id)
        => await _adminService.DeleteRestaurantAsync(id) ? NoContent() : NotFound();

    [HttpPost("restaurants/{id}/pause")]
    [RequiresScope(ApiKeyScopes.Locations, ApiKeyScopes.Write)]
    public async Task<IActionResult> PauseBookings(int id, [FromBody] PauseRestaurantRequest req)
    {
        bool success = await _adminService.PauseRestaurantBookingsAsync(id, req.Minutes);
        return success ? Ok(new MessageResponse { Message = "Bookings paused successfully." }) : NotFound();
    }

    [HttpPost("restaurants/{id}/unpause")]
    [RequiresScope(ApiKeyScopes.Locations, ApiKeyScopes.Write)]
    public async Task<IActionResult> UnpauseBookings(int id)
    {
        bool success = await _adminService.UnpauseRestaurantBookingsAsync(id);
        return success ? Ok(new MessageResponse { Message = "Bookings unpaused successfully." }) : NotFound();
    }

    [HttpPost("restaurants/{id}/extend")]
    [RequiresScope(ApiKeyScopes.Locations, ApiKeyScopes.Write)]
    public async Task<IActionResult> ExtendBookings(int id, [FromBody] ExtendRestaurantRequest req)
    {
        List<BookingDetailDto>? extendedBookings = await _adminService.ExtendAllActiveBookingsAsync(id, req.Minutes);
        return extendedBookings != null
            ? Ok(new { Message = "Bookings extended successfully.", ExtendedBookings = extendedBookings })
            : NotFound();
    }

    [HttpGet("restaurants")]
    [RequiresScope(ApiKeyScopes.Locations, ApiKeyScopes.Read)]
    public async Task<IActionResult> GetRestaurants()
    {
        List<LookupDto> restaurants = await _adminService.GetRestaurantsAsync();
        return Ok(restaurants);
    }

    [HttpGet("restaurants/{restaurantId}/sections")]
    [RequiresScope(ApiKeyScopes.Tables, ApiKeyScopes.Read)]
    public async Task<IActionResult> GetSections(int restaurantId)
    {
        List<LookupDto> sections = await _adminService.GetSectionsAsync(restaurantId);
        return Ok(sections);
    }

    // Scoped under locations (not tables) — this reorders a restaurant's section list, which the
    // issue's resource table groups with the other restaurant/location-shape mutations.
    [HttpPatch("restaurants/{id}/sections/reorder")]
    [RequiresScope(ApiKeyScopes.Locations, ApiKeyScopes.Write)]
    public async Task<IActionResult> ReorderSections(int id, [FromBody] ReorderSectionsRequest req)
    {
        bool? result = await _adminService.ReorderSectionsAsync(id, req.SectionIds);
        return result switch
        {
            null => NotFound(),
            false => BadRequest(new MessageResponse { Message = "sectionIds must include exactly the restaurant's current sections, with no duplicates.", Code = ErrorCodes.RestaurantSectionIdsMismatch }),
            true => NoContent(),
        };
    }

    [HttpGet("restaurants/{restaurantId}/tables")]
    [RequiresScope(ApiKeyScopes.Tables, ApiKeyScopes.Read)]
    public async Task<IActionResult> GetTables(int restaurantId)
    {
        List<SectionDto>? result = await _adminService.GetTablesAsync(restaurantId);
        if (result == null)
        {
            throw new NotFoundException("Restaurant not found or has no sections.") { Code = ErrorCodes.RestaurantNoSections };
        }
        return Ok(result);
    }

    [HttpPost("bookings/{id}/email")]
    [RequiresScope(ApiKeyScopes.Bookings, ApiKeyScopes.Write)]
    public async Task<IActionResult> SendEmail(int id, [FromBody] SendBookingEmailRequest req)
    {
        // Intentionally keeps its catch: SMTP/transport failures (SmtpException,
        // InfrastructureException, etc.) surface here from the email stack and are
        // wrapped as a user-facing 400 "Failed to send: ..." rather than a 500 —
        // this is a deliberate UX choice, not something GlobalExceptionHandler should
        // own (those failures are not domain exceptions).
        try
        {
            SendBookingEmailResult result = await _adminService.SendBookingEmailAsync(id, req);
            return result.Status switch
            {
                SendBookingEmailStatus.NotFound => NotFound(),
                SendBookingEmailStatus.MissingFields => BadRequest(new MessageResponse { Message = "Subject and body are required.", Code = ErrorCodes.BookingEmailFieldsRequired }),
                SendBookingEmailStatus.NoCustomerEmail => BadRequest(new MessageResponse { Message = "Customer email is not available.", Code = ErrorCodes.BookingNoCustomerEmail }),
                _ => Ok(new MessageResponse { Message = $"Email sent to {result.Recipient}." })
            };
        }
        catch (Exception ex)
        {
            return BadRequest(new MessageResponse
            {
                Message = $"Failed to send: {ex.Message}",
                Code = ErrorCodes.BookingEmailSendFailed,
                Args = new Dictionary<string, object> { ["detail"] = ex.Message }
            });
        }
    }

    [HttpPost("bookings/{id}/restore")]
    [RequiresScope(ApiKeyScopes.Bookings, ApiKeyScopes.Write)]
    public async Task<IActionResult> RestoreBooking(int id)
    {
        // BusinessRuleException (booking already active) → 400 is mapped by GlobalExceptionHandler.
        BookingDetailDto? result = await _adminService.RestoreBookingAsync(id);
        return result == null ? NotFound() : Ok(new MessageResponse { Message = "Booking restored successfully." });
    }

    [HttpPut("bookings/{id}")]
    [RequiresScope(ApiKeyScopes.Bookings, ApiKeyScopes.Write)]
    public async Task<IActionResult> AdminUpdateBooking(int id, [FromBody] AdminUpdateBookingRequest req)
    {
        // ValidationException (bad restaurant/table) and BusinessRuleException
        // (update-conflict / seats) → 400 are mapped by GlobalExceptionHandler.
        BookingDetailDto? result = await _adminService.AdminUpdateBookingAsync(id, req);
        return result == null ? NotFound() : Ok(result);
    }
}